+++ crates/alkahest/src/gui/mod.rs
@@ -21,6 +21,7 @@ mod menu;
 mod node_gizmos;
 mod outliner;
 mod profiler;
+pub(crate) mod updater;
 mod util;
 
 pub use configuration::SelectionGizmoMode;